Correcion y optimizacion de las practicas sencillas
