+++ Practics/Practicas sencillas/Codigos/5- gestor tareas.py
@@ -20,7 +20,9 @@ def eliminar_tarea():
     else:
         print("Índice fuera de rango. No se eliminó ninguna tarea.")
 
-while True:
+opcion = ''
+
+while opcion != '4':
     print("\nGestor de Tareas")
     print("1. Agregar Tarea")
     print("2. Ver Tareas")
@@ -35,7 +37,5 @@ while True:
         ver_tareas()
     elif opcion == '3':
         eliminar_tarea()
-    elif opcion == '4':
-        break
-    else:
+    elif opcion != '4':
         print("Opción no válida. Inténtalo de nuevo.")
