+++ Practics/Practicas sencillas/Guias/1- practica adivina el numero.md
@@ -15,11 +15,12 @@ Se importa el módulo `random` para generar números aleatorios, lo cual es nece
 
 ```python
 numero_secreto = random.randint(1, 100)
-intentos = 0
+intento = intentos = 0
 ```
 
 - Se utiliza `random.randint(1, 100)` para generar un número aleatorio entre 1 y 100 y se almacena en la variable `numero_secreto`.
 - La variable `intentos` se inicializa en 0 para llevar un registro de cuántos intentos ha realizado el usuario.
+- La variable `intento` se iguala a la variable `intentos` para utilizar el ciclo while siempre y cuando sea diferente a la variable `numero_secreto`.
 
 **Paso 3:** Mensaje de bienvenida.
 
@@ -32,10 +33,10 @@ Se imprime un mensaje de bienvenida para informar al usuario sobre el tipo de ju
 **Paso 4:** Bucle while principal.
 
 ```python
-while True:
+while intento != numero_secreto:
 ```
 
-Se crea un bucle `while` que se ejecutará indefinidamente hasta que el usuario adivine el número secreto y rompa el bucle con la instrucción `break`.
+Se crea un bucle `while` que se ejecutará indefinidamente hasta que el usuario adivine el número secreto, donde dejara de cumplirse la condición dentro del ciclo por lo que dejara de ejecutarse y seguirá al siguiente paso.
 
 **Paso 5:** Captura del intento del usuario.
 
@@ -50,17 +51,16 @@ Se crea un bucle `while` que se ejecutará indefinidamente hasta que el usuario 
 **Paso 6:** Comparación y mensajes.
 
 ```python
-    if intento == numero_secreto:
-        print("¡Correcto! ¡Adivinaste el número secreto (", numero_secreto, ") en", intentos, "intentos!")
-        break
-    elif intento < numero_secreto:
+    if intento < numero_secreto:
         print("El número secreto es mayor.")
     else:
         print("El número secreto es menor.")
+
+print("¡Correcto! ¡Adivinaste el número secreto (", numero_secreto, ") en", intentos, "intentos!")
 ```
 
 - Se compara el número ingresado por el usuario (`intento`) con el número secreto (`numero_secreto`).
-- Si son iguales, se muestra un mensaje de felicitación, se revela el número secreto y se muestra el número de intentos. Luego, el bucle se rompe con `break`.
+- Si son iguales, el ciclo deja de ejecutarse pasando a la siguiente instrucción donde se imprime el numero secreto y el total de intentos.
 - Si el `intento` es menor que el `numero_secreto`, se informa al usuario que el número secreto es mayor.
 - Si el `intento` es mayor que el `numero_secreto`, se informa al usuario que el número secreto es menor.
