+++ Practics/Practicas sencillas/Guias/5- practica gestor de tareas.md
@@ -1,6 +1,8 @@
 ## Gestor de tareas
 #### Paso 1: Creación del archivo "gestor_de_tareas.py".
 
+- Crea un archivo llamado `gestor_de_tareas.py`
+
 #### Paso 2: Definición de funciones para el gestor de tareas:
 
 - Se crea una lista vacía llamada `tareas` para almacenar las tareas pendientes.
@@ -43,10 +45,11 @@ def eliminar_tarea():
 
 #### Paso 2 (continuación): Bucle principal del gestor de tareas:
 
-- Se inicia un bucle `while` que continúa hasta que el usuario elija salir (`opcion == '4'`).
+- se define la variable de control `opcion = ''` para poder controlar el ciclo while
+- Se inicia un bucle `while` que continúa hasta que el usuario elija salir (`opcion 4`).
 
 ```python
-while True:
+while opcion != '4':
 ```
 
 - Se muestra el menú principal del gestor de tareas con opciones para agregar, ver, eliminar tareas y salir.
@@ -84,19 +87,14 @@ while True:
         eliminar_tarea()
 ```
 
-- Si el usuario elige la opción '4', se rompe el bucle `while` para salir del programa.
+- Si el usuario ingresa una opción diferente a 4, se muestra un mensaje de error, no afecta si se escoge una de las opciones validas.
 
 ```python
-    elif opcion == '4':
-        break
-```
-
-- Si el usuario ingresa una opción no válida, se muestra un mensaje de error.
-
-```python
-    else:
+    elif opcion != '4':
         print("Opción no válida. Inténtalo de nuevo.")
 ```
+
+- Si el usuario elige la opción '4', se deja de cumplir la condición del ciclo `while` y termina el programa.
 
 #### Paso 3: Ejecuta el archivo y gestiona tus tareas utilizando el gestor de tareas.
